The block contextual menu is now separated from the block code itself
